feat: room 增加uuid欄位

+++ app/seq/api/server.go
@@ -21,7 +21,7 @@ func NewServer(c *conf.Config) (*grpc.Server, error) {
 		return nil, err
 	}
 	for _, v := range seqs {
-		bs[v.Id] = &v
+		bs[int64(v.Id)] = &v
 	}
 	pb.RegisterSeqServer(srv, &rpcServer{
 		bs: bs,
